Create README.md
# Hello_FX_GradleKT0

The app is a Basic "Hello World" for JavaFX + FXGL (plus the needed Gluon dependencies)

The Source is KOTLIN
The DSL is KOTLIN

The entire source is in two files to keep things simple, the Cannon App is tested and working for me.  

I spent considerable time searching before I found examples for javaFX that work with the latest Gradle, many deprecation messages, obviously the Kotlin DSL is evolving rapidly.
 
Gradle7 is highly integrated with IntelliJ, for example in the Configuration block, for example if you want to 
see a list of Java Versions just position the cursor to the end of JavaVersion and type '.' and you can pick 
from a list.  (Groovy was pretty nice, but Kotlin is NAILING the it for conciseness, and developer friendliness)